+++ client/src/app/components/create-questions-form/create-questions-form.component.css
@@ -1,0 +1,66 @@
+.bold {
+  font-weight: 500;
+}
+
+h1 {
+  margin: 30px 12% 30px 12%;
+}
+
+.block {
+  display: block;
+}
+
+.item-to-center {
+  margin: auto 12% auto 12%;
+}
+
+h2, h3 {
+  margin-bottom: 20px;
+}
+
+.question-heading {
+  display: flex;
+  justify-content: space-between;
+}
+
+.question-shift {
+  margin-left: 6%;
+}
+
+.required-field {
+  display: flex;
+  align-items: center;
+  margin-bottom: 20px;
+}
+
+h3 {
+  margin-left: 6%;
+}
+
+.choice-entry {
+  display: flex;
+  align-items: center;
+}
+
+.choice-entry > .choice-input {
+  flex-grow: 1;
+}
+
+.choice-shift {
+  margin-left: 12%;
+}
+
+.remove-choice {
+  margin-bottom: 20px;
+  margin-left: 10px;
+}
+
+.add-choice {
+  margin-bottom: 20px;
+}
+
+#add-question {
+  display: flex;
+  justify-content: center;
+  margin-bottom: 20px;
+}
