+++ client/src/app/components/create-questions-form/create-questions-form.component.spec.ts
@@ -1,0 +1,21 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { CreateQuestionsFormComponent } from './create-questions-form.component';
+
+describe('CreateQuestionsFormComponent', () => {
+  let component: CreateQuestionsFormComponent;
+  let fixture: ComponentFixture<CreateQuestionsFormComponent>;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      declarations: [CreateQuestionsFormComponent]
+    });
+    fixture = TestBed.createComponent(CreateQuestionsFormComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
